refactor: use 1 file for image methods
