+++ litellm/llms/custom_httpx/aiohttp_handler.py
@@ -102,7 +102,7 @@ class BaseLLMAIOHTTPHandler:
         api_base: str,
         headers: dict,
         data: dict,
-        timeout: Union[float, httpx.Timeout],
+        timeout: Optional[Union[float, httpx.Timeout]],
         litellm_params: dict,
         stream: bool = False,
         files: Optional[dict] = None,
@@ -458,7 +458,7 @@ class BaseLLMAIOHTTPHandler:
         api_key: str,
         model: Optional[str],
         image: FileTypes,
-        timeout: float,
+        timeout: Optional[float],
         custom_llm_provider: str,
         logging_obj: LiteLLMLoggingObj,
         optional_params: dict,
